All Docstrings

Types

LyoPronto.PrimaryDryFit — Type

PrimaryDryFit: a type for storing experimental data and indicating how it should be fit.

Provided constructors:

PrimaryDryFit(t_Tf, Tfs, Tf_iend, t_Tvw, Tvws, Tvw_iend, t_end)
PrimaryDryFit(t_Tf, Tf::V) where V<:AbstractVector
PrimaryDryFit(t_Tf, Tfs::T) where T<:Tuple
PrimaryDryFit(t_Tf, Tfs, t_end::T) where T<:Unitful.Time
PrimaryDryFit(t_Tf, Tfs, Tvw_end::T) where T<:Unitful.Temperature
PrimaryDryFit(t_Tf, Tfs, Tvw_end::T1, t_end::T2) where {T1<:Unitful.Temperature, T2<:Unitful.Time}
PrimaryDryFit(t_Tf, Tfs, t_Tvw, Tvws) 
PrimaryDryFit(t_Tf, Tfs, t_Tvw, Tvws, t_end)

The use of this struct is determined in large part by the implementation of LyoPronto.obj_expT. If a given field is not available, set it to missing and things should basically work. At least t_Tf and Tfs are expected to always be provided.

With the exception of the two-argument (t_Tf, Tf) constructor, Tfs and Tvws should always be tuples of vectors (one vector per time series).

Tf_iend and Tvw_iend default to [length(Tf) for Tf in Tfs] and [length(Tvw) for Tvw in Tvws], respectively, with one value for each temperature series; they are used to dictate if a given temperature series should be truncated sooner than the full length in the fitting procedure. This implies that all the Tf temperature series are valid initially at the same time points, then stop having measured values after a different number of measurements.

t_end indicates an end of drying, particularly if taken from other measurements (e.g. from Pirani-CM convergence). If set to missing, it is ignored in the objective function.

Principal Cases:

Conventional: provide only t_Tf, Tfs

RF with measured vial wall: provide t_Tf, Tfs, t_Tvw, Tvws,

RF, matching model Tvw to experimental Tf[end] without measured vial wall: provide t_Tf, Tfs, Tvws, set t_Tvw to missing

LyoPronto.RampedVariable — Type

A convenience type for computing temperatures, pressures, etc. with multiple setpoints in sequence, and linear interpolation according to a fixed ramp rate between set points

Three main constructors are available: For a non-varying value, call with one argument:

RampedVariable(constant_setpt)

For one ramp from initial value to set point with indefinite hold, call with two arguments:

RampedVariable(setpts, ramprate)

And for multiple setpoints, call with three arguments:

RampedVariable(setpts, ramprates, holds)

With three arguments, setpts, ramprates, and holds should all be vectors, with lengths N+1, N, N-1 respectively.

The resulting RampedVariable rv = RampedVariable(...) can be called as rv(x) at any (dimensionally consistent) value of x, and will return the value at that time point along the ramp process.

A plot recipe is also provided for this type, e.g. plot(rv; tmax=10u"hr") where tmax indicates where to stop drawing the last setpoint hold.

LyoPronto.RpFormFit — Type

A convenience type for dealing with the common functional form given to Rp and Kv.

An object Rp = RpFormFit(A, B, C) can be called as Rp(x), which simply computes A + B*x/(1 + C*x). Likewise, Kv = RpFormFit(Kc, Kp, Kd) can be called as Kv(p) to get Kc + Kp*p/(1 + Kd*p).

Be careful to pass dimensionally consistent values.

Parameter Fitting

LyoPronto.gen_sol_conv_dim — Method

gen_sol_conv_dim(KRp_prm, otherparams, u0, tspan; kwargs...)

Solve the Pikal model for primary drying with given Kv & Rp, returning the solution object and the set of parameters passed to solve.

otherparams contains: (hf0, c_solid, ρ_solution, Av, Ap, pch, Tsh)

u0 is given as floats (not Unitful quantities), with dimensions [cm, K].

KRp_prm has the form [Kv, R0, A1, A2]; this function assigns the units [W/m^2/K, cm^2*Torr*hr/g, cm*Torr*hr/g, 1/cm] to those numbers

kwargs is passed directly (as is) to the ODE solve call.

This is used as a helper for obj_tT_conv to assemble a function which takes [Kv, R0, A1, A2] and returns sum squared error against experiment. To that end, use this to make an anonymous function of the form gen_sol = x->gen_sol_conv_dim(x, case_params, case_u0, case_tspan). That function is what you will pass to obj_tT_conv.

LyoPronto.gen_sol_rf_dim — Method

gen_sol_rf_dim(fitprm, params_bunch, u0, tspan; kwargs...)

Solve the lumped-capacitance model for microwave-assisted primary drying with given fit parameters, returning the solution object and the set of parameters passed to solve.

fitprm has the form [α, Kvwf, Bf, Bvw]; this function assigns the units [cm^1.5, cal/s/K/cm^2, Ω/m^2, Ω/m^2] to those numbers.

params_bunch contains a full listing of parameters used for the model, according to lumped_cap_rf, including dummy values for the fit parameters.

u0 is given as floats (not Unitful quantities), with dimensions [g, K, K].

kwargs is passed directly (as is) to the ODE solve call.

This is used as a helper for obj_tT_rf, obj_tTT_rf, obj_ttTT_rf to assemble a function which takes [α, Kvwf, Bf, Bvw] and returns sum squared error against experiment. To that end, use this to make an anonymous function of the form gen_sol = x->gen_sol_rf_dim(x, case_params, case_u0, case_tspan). That function is what you will pass to obj_tT_rf (or similar).

LyoPronto.obj_expT — Method

obj_expT(sol, pdfit; tweight=1, verbose=true, rf = true)

Experimental (in the software engineering sense)! Evaluate an objective function which compares model solution computed by sol to experimental data in pdfit.

sol is a solution to an appropriate model; see gen_sol_conv_dim and gen_sol_rf_dim for some helper functions for this.

pdfit is an instance of PrimaryDryFit, which contains some information about what to compare.

tweight = 1 gives the weighting (in K^2/hr^2) of the total drying time in the objective, as compared to the temperature error.

Note that if pdfit has vial wall temperatures (i.e. ismissing(pdfit.Tvws) == false), the third-index variable returned by gen_sol is assumed to be temperature, as is true for gen_sol_rf_dim.

If there are multiple series of Tf in pdfit, squared error is computed for each separately then summed; likewise for Tvw.

I've considered writing several methods and dispatching on pdfit somehow, which would be cool and might individually be easier to read. But control flow might be harder to document and explain, and this should work just fine.

LyoPronto.obj_tTT_rf — Method

obj_tTT_rf(fitprm, gen_sol, tTTdat; t_end=0.0u"hr", tweight=1, verbose=true)

Evaluate an objective function which compares model solution with fitprm to experimental data in tTTdat.

gen_sol is a function taking [α, Kvwf, Bf, Bvw] and returning a solution to lumped-capacitance microwave-assisted model; see gen_sol_rf_dim.

tTTdat is experimental temperature series, of the form (time, Tf, Tvw), so with frozen and vial wall temperatures taken at the same time points. See also obj_tT_rf and obj_ttTT_rf.

tweight gives the weighting (in K^2/hr^2) of the end of drying in the objective, as compared to the temperature error.

t_end has a default value of 0.0u"hr", which (if left at default) is not included in the objective function.

t_end defaults to missing, in which case it is excluded from the objective. If another value is passed, then the model drying time is compared to that value and included in the objective.

LyoPronto.obj_tT_conv — Method

obj_tT_conv(KRp_prm, gen_sol, tTdat; t_end=0.0u"hr", tweight=1, verbose = true)

Evaluate an objective function which compares model solution with KRp_prm to experimental data in tTdat.

Arguments:

gen_sol is a function taking [Kv, R0, A1, A2] and returning a solution to Pikal model; see gen_sol_conv_dim.

tTdat is experimental temperature series, of the form (time, Tf).

tweight gives the weighting (in K^2/hr^2) of the end of drying in the objective, as compared to the temperature error.

t_end defaults to missing, in which case it is excluded from the objective. If another value is passed, then the model drying time is compared to that value and included in the objective.

verbose defaults to true, in which case each call to this function prints info on the passed parameters, etc.

LyoPronto.obj_tT_rf — Method

obj_tT_rf(fitprm, gen_sol, tTdat; t_end=0.0u"hr", tweight=1, Tvw_end = 0.0u"K", verbose=true)

Evaluate an objective function which compares model solution with fitprm to experimental data in tTdat.

gen_sol is a function taking [α, Kvwf, Bf, Bvw] and returning a solution to lumped-capacitance microwave-assisted model; see gen_sol_rf_dim.

tTTdat is experimental temperature series, of the form (time, Tf), so with frozen temperatures only. See also obj_tTT_rf and obj_ttTT_rf.

Tvw_end defaults to missing, in which case vial wall temperatures are excluded from the objective. If another value is passed, then the final model vial wall temperature is compared to that value and included in the objective.

tweight gives the weighting (in K^2/hr^2) of the end of drying in the objective, as compared to the temperature error.

t_end defaults to missing, in which case it is excluded from the objective. If another value is passed, then the model drying time is compared to that value and included in the objective.

LyoPronto.obj_ttTT_rf — Method

obj_ttTT_rf(fitprm, gen_sol, ttTTdat; t_end=missing, tweight=1, verbose=true)

Evaluate an objective function which compares model solution with fitprm to experimental data in tTdat.

gen_sol is a function taking [α, Kvwf, Bf, Bvw] and returning a solution to lumped-capacitance microwave-assisted model; see gen_sol_rf_dim.

ttTTdat is experimental temperature series, of the form (time_Tf, time_vw, Tf, Tvw), so with Tf and Tvw having separate time points. This is useful if there is an early temperature rise in Tf, but Tvw continues to be reliable, so the model can fit to as much of Tvw as reasonable. See also obj_tTT_rf and obj_tT_rf.

tweight gives the weighting (in K^2/hr^2) of the end of drying in the objective, as compared to the temperature error.

t_end defaults to missing, in which case it is excluded from the objective. If another value is passed, then the model drying time is compared to that value and included in the objective.

Plot Recipes

LyoPronto.exptfplot — Function

exptfplot(time, T1, [T2, ...])
exptfplot!(time, T1, [T2, ...])

Plot recipe for one or more experimentally measured product temperatures, all at same times. This recipe adds one series for each passed temperature series, so pass labels as appropriate.

LyoPronto.exptfplot! — Function

exptfplot(time, T1, [T2, ...])
exptfplot!(time, T1, [T2, ...])

Plot recipe for one or more experimentally measured product temperatures, all at same times. This recipe adds one series for each passed temperature series, so pass labels as appropriate.

LyoPronto.exptvwplot — Function

exptvwplot(time, temperature; trim)
exptvwplot!(time, temperature; trim)

Plot recipe for a set of experimentally measured vial wall temperatures. This recipe adds only one series to the plot. trim is an integer, indicating how many points to skip at a time, so that the dotted line looks dotted even with noisy data.

LyoPronto.exptvwplot! — Function

exptvwplot(time, temperature; trim)
exptvwplot!(time, temperature; trim)

Plot recipe for a set of experimentally measured vial wall temperatures. This recipe adds only one series to the plot. trim is an integer, indicating how many points to skip at a time, so that the dotted line looks dotted even with noisy data.

LyoPronto.modconvtplot — Function

modconvtplot(sols)
modconvtplot!(sols)

Plot recipe for one or multiple solutions to the Pikal model, e.g. the output of gen_sol_conv_dim. This adds one series to the plot for each passed solution, so pass as many labels (e.g. ["Tf1" "Tf2"]) to this plot call as solutions to add labels to the legend.

LyoPronto.modconvtplot! — Function

modconvtplot(sols)
modconvtplot!(sols)

Plot recipe for one or multiple solutions to the Pikal model, e.g. the output of gen_sol_conv_dim. This adds one series to the plot for each passed solution, so pass as many labels (e.g. ["Tf1" "Tf2"]) to this plot call as solutions to add labels to the legend.

LyoPronto.modrftplot — Function

modrftplot(sol)
modrftplot!(sol)

Plot recipe for one solution to the lumped capacitance model, e.g. the output of gen_sol_rf_dim. This adds two series to the plot, so pass two labels (e.g. ["Tf" "Tvw"]) to this plot call to add labels to the legend.

LyoPronto.modrftplot! — Function

modrftplot(sol)
modrftplot!(sol)

Plot recipe for one solution to the lumped capacitance model, e.g. the output of gen_sol_rf_dim. This adds two series to the plot, so pass two labels (e.g. ["Tf" "Tvw"]) to this plot call to add labels to the legend.

LyoPronto.qrf_integrate — Method

qrf_integrate(sol, RF_params)

Compute the integral over time of each heat transfer mode in the lumped capacitance model.

Returns as a Dict{String, Quantity{...}}, with string keys Qsub, Qshf, Qvwf, QRFf, QRFvw.

LyoPronto.tendplot — Function

tendplot(t_end)
tendplot!(t_end)

Plot recipe that adds a labeled vertical line to the plot at time t_end. A default label and styling are applied, but these can be modified by keyword arguments as usual

LyoPronto.tendplot! — Function

tendplot(t_end)
tendplot!(t_end)

Plot recipe that adds a labeled vertical line to the plot at time t_end. A default label and styling are applied, but these can be modified by keyword arguments as usual

Vial Dimensions

LyoPronto.get_vial_mass — Method

get_vial_mass(vialsize::String)

Return vial mass for given ISO vial size.

Uses a table provided by Schott, stored internally in a CSV.

LyoPronto.get_vial_radii — Method

get_vial_radii(vialsize::String)

Return inner and outer radius for passed ISO vial size.

Uses a table provided by Schott, stored internally in a CSV.

LyoPronto.get_vial_shape — Method

get_vial_shape(vialsize::String)

Return a Dict{Symbol, Any} with a slew of vial dimensions, useful for drawing the shape of the vial with make_outlines.

LyoPronto.get_vial_thickness — Method

get_vial_thickness(vialsize::String)

Return vial wall thickness for given ISO vial size.

Uses a table provided by Schott, stored internally in a CSV.

LyoPronto.make_outlines — Method

make_outlines(dims, Vfill)

Return a sequence of points (ready to be made into Plots.Shapes for the vial and fill volume, with Unitful dimensions, for given vial dimensions.

This is a convenience function for making figures illustrating fill depth.

Model Equations

LyoPronto.end_drying_callback — Constant

A callback for use in simulating either the Pikal or RF model.

Terminates the time integration when end_cond evaluates to true.

LyoPronto.end_cond — Method

end_cond(u, t, integ)

Compute the end condition for primary drying (that mf or hf approaches zero).

LyoPronto.lumped_cap_rf — Method

lumped_cap_rf(u, params, tn)

Compute the right-hand-side function for the ODEs making up the lumped-capacitance microwave-assisted model.

Specifically, this is [dmf/dt, dTf/dt, dTvw/dt] given u = [mf, Tf, Tvw]. u is taken without units but assumed to have the units of [g, K, K] (which is internally added). tn is assumed to be in hours (internally added), so dudt is returned with assumed units [g/hr, K/hr, K/hr] to be consistent.

The full set of necessary parameters is given in the form of a tuple-of-tuples:

params = (   
    (Rp, h_f0, cSolid, ρ_solution),
    (K_shf_f, A_v, A_p),
    (pch, Tsh, P_per_vial),
    (m_f0, cp_f, m_v, cp_v, A_rad),
    (f_RF, epp_f, epp_vw),
    (K_vwf, B_f, B_vw, alpha),
)

These should all be Unitful quantities with appropriate dimensions, with some exceptions which are callables returning quantities. See RpFormFit and RampedVariable for convenience types that can help with these cases.

Rp(x) with x a length returns mass transfer resistance (as a Unitful quantity)

K_shf_f(p) with p a pressure returns heat transfer coefficient (as a Unitful quantity).

Tsh(t), pch(t), P_per_vial(t) return shelf temperature, chamber pressure, and microwave power respectively at time t.

A_rad and alpha are used only in the LC2 and LC3 versions of the model, and can be left out.

For implementation details, see lumped_cap_rf_LC1.

LyoPronto.lumped_cap_rf_LC1 — Method

lumped_cap_rf_LC1(u, params, tn)

This does the work for lumped_cap_rf, but returns dudt, [Q_sub, Q_shf, Q_vwf, Q_RF_f, Q_RF_vw, Q_shw] with Q_... as Unitful quantities in watts. The extra results are helpful in investigating the significance of the various heat transfer modes, but are not necessary in the ODE integration.

LC1: Qshw evaluated with Kshf; shape factor included; α=0 My preferred version.

LyoPronto.lyo_1d_dae! — Method

lyo_1d_dae!(du, u, params, t)

Internal implementation of the Pikal model. See lyo_1d_dae_f for the wrapped version, which is more fully documented.

LyoPronto.lyo_1d_dae_f — Function

lyo_1d_dae_f = ODEFunction(lyo_1d_dae!, mass_matrix=lyo_1d_mm)

Compute the right hand side function for the Pikal model.

The DAE system which is the Pikal model (1 ODE, one nonlinear algebraic equation for pseudosteady conditions) is here treated as a constant-mass-matrix implicit ODE system. The implementation is in lyo_1d_dae!

The initial conditions u0 = [h_f, Tf] should be unitless, but are internally assigned to be in [cm, K]. The unitless time is taken to be in hours, so derivatives are given in unitless [cm/hr, K/hr].

params has the form:

params = (
    (Rp, hf0, c_solid, ρ_solution),
    (Kshf, Av, Ap),
    (pch, Tsh) ,
)

where those listed following are callables returning Quantitys, and the rest are Quantitys. See RpFormFit and RampedVariable for convenience types that can help with the callables.

Rp(x) with x a length returns mass transfer resistance (as a Unitful quantity)

Kshf(p) with p a pressure returns heat transfer coefficient (as a Unitful quantity).

Tsh(t), pch(t) return shelf temperature and chamber pressure respectively at time t.

Physical Properties

LyoPronto.calc_psub — Method

calc_psub(T::F) where F<:Number
calc_psub(T::Q) where Q<:Quantity

Compute pressure (in Pascals) of sublimation at temperature T in Kelvin.

This is essentially an Arrhenius fit, where we compute: psub = pref * exp(-ΔHsub*Mw/R/T)

LyoPronto.Dielectric — Module

Single-purpose module for computing the dielectric loss coefficient of ice, as a function of temperature and frequency. This module exports a function ϵpp_f(T, f) for doing so. (Set as a separate module just to keep namespaces clean.) This code is a nearly-direct implementation of the correlation, Eqs. 3-6 presented in: Takeshi Matsuoka, Shuji Fujita, Shinji Mae; Effect of temperature on dielectric properties of ice in the range 5–39 GHz. J. Appl. Phys. 15 November 1996; 80 (10): 5884–5890. https://doi.org/10.1063/1.363582