LyoPronto.calc_psub — Method

calc_psub(T)

Compute pressure (in Pascals) of sublimation at temperature T in Kelvin.

From...somewhere else. Used in LyoPRONTO

LyoPronto.gen_sol_conv_dim — Method

obj_tT_conv(KRp_prm, otherparams, u0, tTdat; tweight=1)

Evaluate an objective function which compares model solution with KRp_prm to experimental data in tTdat.

otherparams contains: (hf0, c_solid, ρ_solution, Av, Ap, pch, Tsh)

u0 is unitless, but with dimensions [cm, K].

tweight gives the weighting (in K^2/hr^2) of the end of drying in the objective, as compared to the temperature error.

LyoPronto.get_vial_radii — Method

get_vial_radii(vialsize::String)

Return inner and outer radius for passed ISO vial size.

Uses a table provided by Schott, stored internally in a CSV.

LyoPronto.get_vial_thickness — Method

get_vial_thickness(vialsize::String)

Return vial wall thickness for given ISO vial size.

Uses a table provided by Schott, stored internally in a CSV.

LyoPronto.lumped_cap_rf — Method

function lumped_cap_rf(du, u, p, t)

TODO: include all parameters TODO: handle dimensionality

LyoPronto.make_outlines — Method

make_outlines(dims, Vfill)

Return Plots.Shapes for vial and fill volume, with Unitful dimensions, for given vial dimensions. Convenience function for making some figures.

LyoPronto.obj_tT_conv — Method

obj_tT_conv(KRp_prm, gen_sol, tTdat; tweight=1)

Evaluate an objective function which compares model solution with KRp_prm to experimental data in tTdat.

tweight gives the weighting (in K^2/hr^2) of the end of drying in the objective, as compared to the temperature error.

LyoPronto.tplotexperimental — Function

tplotexperimental(time, T1, [T2, ...])

Plot recipe for one or more experimentally measured product temperatures, all at same times.

LyoPronto.tplotexpvw — Function

tplotexpvw(time, temperature)

Plot recipe for a set of experimentally measured vial wall temperatures.