All Docstrings

Types

LyoPronto.RampedVariable — Type

A convenience type for computing temperatures, pressures, etc. with multiple setpoints in sequence, and linear interpolation according to a fixed ramp rate between set points

Three main constructors are available: For a non-varying value, call with one argument: RampedVariable(constant_setpt) For one ramp from initial value to set point with indefinite hold, call with two arguments: RampedVariable(setpts, ramprate) And for multiple setpoints, call with three arguments: RampedVariable(setpts, ramprates, holds)

With three arguments, setpts, ramprates, and holds should all be vectors, with lengths N+1, N, N-1 respectively.

The resulting RampedVariable rv = RampedVariable(...) can be called as rv(x) at any value of x, and will return the value at that time point along the ramp process.

LyoPronto.RpFormFit — Type

A convenience type for dealing with the common functional form given to Rp and Kv.

An object Rp = RpFormFit(A, B, C) can be called as Rp(x), which simply computes A + B*x/(1 + C*x). Likewise, Kv = RpFormFit(Kc, Kp, Kd) can be called as Kv(p) to get Kc + Kp*p/(1 + Kd*p).

Parameter Fitting

LyoPronto.gen_sol_conv_dim — Method

gen_sol_conv_dim(KRp_prm, otherparams, u0, tspan; kwargs...)

Solve the Pikal model for primary drying with given Kv & Rp, returning the solution object and the set of parameters passed to solve.

otherparams contains: (hf0, c_solid, ρ_solution, Av, Ap, pch, Tsh)

u0 is given as floats (not Unitful quantities), with dimensions [cm, K].

KRp_prm has the form [Kv, R0, A1, A2]; this function assigns the units [W/m^2/K, cm^2Torrhr/g, cmTorrhr/g, 1/cm] to those numbers

kwargs is passed directly (as is) to the ODE solve call.

This is used as a helper for obj_tT_conv to assemble a function which takes [Kv, R0, A1, A2] and returns sum squared error against experiment. To that end, use this to make an anonymous function of the form gen_sol = x->gen_sol_conv_dim(x, case_params, case_u0, case_tspan). That function is what you will pass to obj_tT_conv.

LyoPronto.gen_sol_rf_dim — Method

gen_sol_rf_dim(fitprm, params_bunch, u0, tspan; kwargs...)

Solve the lumped-capacitance model for microwave-assisted primary drying with given fit parameters, returning the solution object and the set of parameters passed to solve.

fitprm has the form [α, Kvwf, Bf, Bvw]; this function assigns the units [cm^1.5, cal/s/K/cm^2, Ω/m^2, Ω/m^2] to those numbers.

params_bunch contains a full listing of parameters used for the model, according to lumped_cap_rf, including dummy values for the fit parameters.

u0 is given as floats (not Unitful quantities), with dimensions [g, K, K].

kwargs is passed directly (as is) to the ODE solve call.

This is used as a helper for obj_tT_rf, obj_tTT_rf, obj_ttTT_rf to assemble a function which takes [α, Kvwf, Bf, Bvw] and returns sum squared error against experiment. To that end, use this to make an anonymous function of the form gen_sol = x->gen_sol_rf_dim(x, case_params, case_u0, case_tspan). That function is what you will pass to obj_tT_rf (or similar).

LyoPronto.obj_tTT_rf — Method

obj_tTT_rf(fitprm, gen_sol, tTTdat; t_end=0.0u"hr", tweight=1, verbose=true)

Evaluate an objective function which compares model solution with fitprm to experimental data in tTTdat.

gen_sol is a function taking [α, Kvwf, Bf, Bvw] and returning a solution to lumped-capacitance microwave-assisted model; see gen_sol_rf_dim.

tTTdat is experimental temperature series, of the form (time, Tf, Tvw), so with frozen and vial wall temperatures taken at the same time points. See also obj_tT_rf and obj_ttTT_rf.

tweight gives the weighting (in K^2/hr^2) of the end of drying in the objective, as compared to the temperature error.

t_end has a default value of 0.0u"hr", which (if left at default) is replaced with the last given time point.

LyoPronto.obj_tT_conv — Method

obj_tT_conv(KRp_prm, gen_sol, tTdat; t_end=0.0u"hr", tweight=1, verbose = true)

Evaluate an objective function which compares model solution with KRp_prm to experimental data in tTdat.

Arguments:

gen_sol is a function taking [Kv, R0, A1, A2] and returning a solution to Pikal model; see gen_sol_conv_dim.

tTdat is experimental temperature series, of the form (time, Tf).

tweight gives the weighting (in K^2/hr^2) of the end of drying in the objective, as compared to the temperature error.

t_end has a default value of 0.0u"hr", which (if left at default) is replaced with the last time point.

verbose defaults to true, in which case each call to this function prints info on the passed parameters, etc.

LyoPronto.obj_tT_rf — Method

obj_tT_rf(fitprm, gen_sol, tTdat; t_end=0.0u"hr", tweight=1, Tvw_end = 0.0u"K", verbose=true)

Evaluate an objective function which compares model solution with fitprm to experimental data in tTdat.

gen_sol is a function taking [α, Kvwf, Bf, Bvw] and returning a solution to lumped-capacitance microwave-assisted model; see gen_sol_rf_dim.

tTTdat is experimental temperature series, of the form (time, Tf), so with frozen temperatures only. See also obj_tTT_rf and obj_ttTT_rf.

Tvw_end defaults to 0.0u"K", in which case vial wall temperatures are excluded from the objective. If another value is passed, then the final model vial wall temperature is compared to that value and included in the objective.

tweight gives the weighting (in K^2/hr^2) of the end of drying in the objective, as compared to the temperature error.

t_end has a default value of 0.0u"hr", which (if left at default) is replaced with the last given time point.

LyoPronto.obj_ttTT_rf — Method

obj_ttTT_rf(fitprm, gen_sol, tTTdat; t_end=0.0u"hr", tweight=1, verbose=true)

Evaluate an objective function which compares model solution with fitprm to experimental data in tTdat.

gen_sol is a function taking [α, Kvwf, Bf, Bvw] and returning a solution to lumped-capacitance microwave-assisted model; see gen_sol_rf_dim.

tTTdat is experimental temperature series, of the form (time_Tf, time_vw, Tf, Tvw), so with Tf and Tvw having separate time points. This is useful if there is an early temperature rise in Tf, but Tvw continues to be reliable, so the model can fit to as much of Tvw as reasonable. See also obj_tTT_rf and obj_tT_rf.

tweight gives the weighting (in K^2/hr^2) of the end of drying in the objective, as compared to the temperature error.

t_end has a default value of 0.0u"hr", which (if left at default) is replaced with the last given time point.

Plot Recipes

LyoPronto.qrf_integrate — Method

qrf_integrate(sol, RF_params)

Compute the integral over time of each heat transfer mode in the lumped capacitance model.

Returns as a Dict{String, Quantity{...}}, with string keys Qsub, Qshf, Qvwf, QRFf, QRFvw.

LyoPronto.tendplot — Function

tendplot(t_end)
tendplot!(t_end)

Plot recipe that adds a labeled vertical line to the plot at time t_end. A default label and styling are applied, but these can be modified by keyword arguments as usual

LyoPronto.tendplot! — Function

tendplot(t_end)
tendplot!(t_end)

Plot recipe that adds a labeled vertical line to the plot at time t_end. A default label and styling are applied, but these can be modified by keyword arguments as usual

LyoPronto.tplotexperimental — Function

tplotexperimental(time, T1, [T2, ...])
tplotexperimental!(time, T1, [T2, ...])

Plot recipe for one or more experimentally measured product temperatures, all at same times. This recipe adds one series for each passed temperature series, so pass labels as appropriate.

LyoPronto.tplotexperimental! — Function

tplotexperimental(time, T1, [T2, ...])
tplotexperimental!(time, T1, [T2, ...])

Plot recipe for one or more experimentally measured product temperatures, all at same times. This recipe adds one series for each passed temperature series, so pass labels as appropriate.

LyoPronto.tplotexpvw — Function

tplotexpvw(time, temperature)
tplotexpvw!(time, temperature)

Plot recipe for a set of experimentally measured vial wall temperatures. This recipe adds only one series to the plot.

LyoPronto.tplotexpvw! — Function

tplotexpvw(time, temperature)
tplotexpvw!(time, temperature)

Plot recipe for a set of experimentally measured vial wall temperatures. This recipe adds only one series to the plot.

LyoPronto.tplotmodelconv — Function

tplotmodelconv(sols)
tplotmodelconv!(sols)

Plot recipe for one or multiple solutions to the Pikal model, e.g. the output of gen_sol_conv_dim. This adds one series to the plot for each passed solution, so pass as many labels (e.g. ["Tf1" "Tf2"]) to this plot call as solutions to add labels to the legend.

LyoPronto.tplotmodelconv! — Function

tplotmodelconv(sols)
tplotmodelconv!(sols)

Plot recipe for one or multiple solutions to the Pikal model, e.g. the output of gen_sol_conv_dim. This adds one series to the plot for each passed solution, so pass as many labels (e.g. ["Tf1" "Tf2"]) to this plot call as solutions to add labels to the legend.

LyoPronto.tplotmodelrf — Function

tplotmodelrf(sol)
tplotmodelrf!(sol)

Plot recipe for one solution to the lumped capacitance model, e.g. the output of gen_sol_rf_dim. This adds two series to the plot, so pass two labels (e.g. ["Tf" "Tvw"]) to this plot call to add labels to the legend.

LyoPronto.tplotmodelrf! — Function

tplotmodelrf(sol)
tplotmodelrf!(sol)

Plot recipe for one solution to the lumped capacitance model, e.g. the output of gen_sol_rf_dim. This adds two series to the plot, so pass two labels (e.g. ["Tf" "Tvw"]) to this plot call to add labels to the legend.

Vial Dimensions

LyoPronto.get_vial_mass — Method

get_vial_mass(vialsize::String)

Return vial mass for given ISO vial size.

Uses a table provided by Schott, stored internally in a CSV.

LyoPronto.get_vial_radii — Method

get_vial_radii(vialsize::String)

Return inner and outer radius for passed ISO vial size.

Uses a table provided by Schott, stored internally in a CSV.

LyoPronto.get_vial_shape — Method

get_vial_shape(vialsize::String)

Return a Dict{Symbol, Any} with a slew of vial dimensions, useful for drawing the shape of the vial with make_outlines.

LyoPronto.get_vial_thickness — Method

get_vial_thickness(vialsize::String)

Return vial wall thickness for given ISO vial size.

Uses a table provided by Schott, stored internally in a CSV.

LyoPronto.make_outlines — Method

make_outlines(dims, Vfill)

Return a sequence of points (ready to be made into Plots.Shapes for the vial and fill volume, with Unitful dimensions, for given vial dimensions.

This is a convenience function for making figures illustrating fill depth.

Model Equations

LyoPronto.end_drying_callback — Constant

A callback for use in simulating either the Pikal or RF model.

Terminates the time integration when end_cond evaluates to true.

LyoPronto.calc_psub — Method

calc_psub(T::F) where F<:Number
calc_psub(T::Q) where Q<:Quantity

Compute pressure (in Pascals) of sublimation at temperature T in Kelvin.

This is essentially an Arrhenius fit, where we compute: psub = pref * exp(-ΔHsub*Mw / R T)

LyoPronto.end_cond — Method

end_cond(u, t, integ)

Compute the end condition for primary drying (that mf or hf approaches zero).

LyoPronto.lumped_cap_rf — Method

lumped_cap_rf(u, params, tn)

Compute the right-hand-side function for the ODEs making up the lumped-capacitance microwave-assisted model.

Specifically, this is [dmf/dt, dTf/dt, dTvw/dt] given u = [mf, Tf, Tvw]. u is taken without units but assumed to have the units of [g, K, K] (which is internally added). tn is assumed to be in hours (internally added), so dudt is returned with units [g/hr, K/hr, K/hr] to be consistent.

The full set of necessary parameters is given in the form of a tuple-of-tuples:

params = (   
    (Rp, h_f0, cSolid, ρ_solution),
    (K_shf_f, A_v, A_p),
    (pch, Tsh, P_per_vial),
    (m_f0, cp_f, m_v, cp_v, A_rad),
    (f_RF, epp_f, epp_vw),
    (alpha, K_vwf, B_f, B_vw),
)

These should all be Unitful quantities with appropriate dimensions, with some exceptions which are callables returning quantities. See RpFormFit and `RampedVariable for convenience types that can help with these cases.

Rp(x) with x a length returns mass transfer resistance (as a Unitful quantity)

K_shf_f(p) with p a pressure returns heat transfer coefficient (as a Unitful quantity).

Tsh(t), pch(t), P_per_vial(t) return shelf temperature, chamber pressure, and microwave power respectively at time t.

For implementation details, see lumped_cap_rf_model.

LyoPronto.lumped_cap_rf_model — Method

lumped_cap_rf_model(u, params, tn)

This does the work for lumped_cap_rf, but returns dudt, [Q_sub, Q_shf, Q_vwf, Q_RF_f, Q_RF_vw, Q_shw] with Q_... as Unitful quantities in watts. The extra results are helpful in investigating the significance of the various heat transfer modes.

LyoPronto.lyo_1d_dae! — Method

lyo_1d_dae!(du, u, params, t)

Internal implementation of the Pikal model. See lyo_1d_dae_f for the wrapped version, which is more fully documented.

LyoPronto.lyo_1d_dae_f — Function

lyo_1d_dae_f = ODEFunction(lyo_1d_dae!, mass_matrix=lyo_1d_mm)

Compute the right hand side function for the Pikal model.

The DAE system which is the Pikal model (1 ODE, one nonlinear algebraic equation for pseudosteady conditions) is here treated as a constant-mass-matrix implicit ODE system. The implementation is in lyo_1d_dae!

The initial conditions u0 = [h_f, Tf] should be unitless, but are internally assigned to be in [cm, K]. The unitless time is taken to be in hours, so derivatives are given in unitless [cm/hr, K/hr].

params has the form:

params = (
    (Rp, hf0, c_solid, ρ_solution),
    (Kshf, Av, Ap),
    (pch, Tsh) ,
)

where some quantities with Unitful units and some are callables returning quantities See RpFormFit and RampedVariable for convenience types that can help with the callables.

Rp(x) with x a length returns mass transfer resistance (as a Unitful quantity)

K_shf_f(p) with p a pressure returns heat transfer coefficient (as a Unitful quantity).

Tsh(t), pch(t) return shelf temperature and chamber pressure respectively at time t.